refactor(authSession): split into session, events, issuer modules
- session.ts: OIDC session factory (WebSession/SessionCore, database backends)
- events.ts: SessionEvents class (pure EventEmitter shim, zero side-effects)
- issuer.ts: issuer discovery via /.well-known/openid-configuration
- authSession.ts: login compatibility shim, legacy event wiring, authSession
  assembly; re-exports authSession and SessionWithLegacyEvents type

All existing imports continue working unchanged — authSession.ts is a
drop-in replacement with the same exports.
